requirements.txt
fastapi
uvicorn[standard]
requests
python-multipart
aiofiles
jinja2
websockets
python-dotenv
